daily-sync-backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/ML Libraries
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-community>=0.0.20
openai>=1.0.0
anthropic>=0.23.0

# Vector Database
chromadb==0.4.22

# Database
aiosqlite==0.19.0

# Audio Processing (for Whisper)
openai-whisper==20231117
pydub==0.25.1

# Image Processing
Pillow==10.2.0

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0
urllib3<2.0.0  # Fix OpenSSL warning on macOS

# Optional: CrewAI (if you want to use it later)
# crewai==0.28.8
